+++ examples/web-dashboard/src/main/scala/com/twitter/web/dashboard/views/UserView.scala
@@ -1,6 +1,6 @@
 package com.twitter.web.dashboard.views
 
-import com.twitter.finatra.response.Mustache
+import com.twitter.finatra.http.response.Mustache
 
 @Mustache("user")
 case class UserView(firstName: String, lastName: String)