+++ jackson/src/main/java/com/twitter/finatra/request/JsonIgnoreBody.java
@@ -3,7 +3,6 @@ package com.twitter.finatra.request;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;